Add clear.js
Added moderation features for Clear.js and cleaned up some functionality

+++ src/util/functions.js
@@ -1,18 +1,31 @@
+import _ from 'lodash'
 import {
     BLOCKED_WORDS,
 } from './constants'
 
-async function checkTwitchChat(userstate, message, channel, client) {
+const main = async (chatBot) => {
+    await chatBot.connect().catch(console.error);
+    chatBot.color('CadetBlue');
+};
+
+const checkTwitchChat = async (userstate, message, channel, client, adjustedUserstate) => {
+    if (adjustedUserstate.mod === 'mod') return;
     message = message.toLocaleLowerCase();
     let shouldSendMessage = false;
-    
     shouldSendMessage = BLOCKED_WORDS.some(blockedWord => message.includes(blockedWord.toLocaleLowerCase()));
+    
     if (shouldSendMessage) {
-        await client.deletemessage(channel, userstate.id)
-        client.say(channel, `${userstate.username}, oh. You said a bad word...`)
-    }
+        await client.deletemessage(channel, userstate.id);
+        client.say(channel, `${userstate.username}, oh. You said a bad word...`);
+    };
+};
+
+function userObj(user) {
+	return _.mapKeys(user, n => _.camelCase(n));
 };
 
 export {
-    checkTwitchChat
-}
+    main,
+    checkTwitchChat,
+    userObj
+};
